requirements.txt
streamlit
psycopg2-binary
bcrypt
cloudinary
Pillow
streamlit-cookies-controller